requirements.txt
annotated-types==0.7.0
anyio==4.5.0
APScheduler==3.11.0
asgiref==3.9.2
attrs==24.2.0
bcrypt==4.3.0
beautifulsoup4==4.12.3
blinker==1.8.2
bs4==0.0.2
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.7
contourpy==1.3.0
cryptography==44.0.3
cycler==0.12.1
Django==5.2.6
django-cors-headers==3.13.0
djangorestframework==3.16.1
dnspython==2.7.0
ecdsa==0.19.1
email_validator==2.2.0
exceptiongroup==1.2.2
fastapi==0.115.12
fastapi-cli==0.0.7
Flask==3.0.3
Flask-APScheduler==1.13.1
Flask-SQLAlchemy==3.1.1
fonttools==4.54.1
h11==0.14.0
httpcore==1.0.5
httptools==0.6.4
httpx==0.27.2
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
joblib==1.4.2
kiwisolver==1.4.7
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.9.2
mdurl==0.1.2
mysql-connector-python==9.3.0
np==1.0.2
numpy==2.1.2
orjson==3.10.18
outcome==1.3.0.post0
package-name==0.1
packaging==24.1
pandas==2.2.3
passlib==1.7.4
pdf2image==1.17.0
pdfminer.six==20250327
pdfplumber==0.11.6
pillow==10.4.0
psycopg2-binary==2.9.10
pyasn1==0.6.1
pycparser==2.22
pydantic==2.11.5
pydantic-extra-types==2.10.5
pydantic-settings==2.9.1
pydantic_core==2.33.2
Pygments==2.19.1
pymongo==4.12.1
PyMySQL==1.1.2
pyparsing==3.2.0
pypdfium2==4.30.1
PySocks==1.7.1
pytesseract==0.3.13
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-jose==3.5.0
python-multipart==0.0.20
python-telegram-bot==21.6
pytz==2025.2
PyYAML==6.0.2
requests==2.32.3
rich==14.0.0
rich-toolkit==0.14.7
rsa==4.9.1
scikit-learn==1.5.2
scipy==1.14.1
selenium==4.25.0
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.6
SQLAlchemy==2.0.40
sqlparse==0.5.3
starlette==0.46.2
threadpoolctl==3.5.0
trio==0.26.2
trio-websocket==0.11.1
typer==0.16.0
typing-inspection==0.4.1
typing_extensions==4.12.2
tzdata==2025.2
tzlocal==5.3.1
ujson==5.10.0
urllib3==2.2.3
uvicorn==0.34.2
watchfiles==1.0.5
websocket-client==1.8.0
websockets==15.0.1
Werkzeug==3.0.4
Wikipedia-API==0.7.1
wordcloud==1.9.3
wsproto==1.2.0
python-socketio==5.11.0
